requirements.txt
Flask==2.1.0
flask_cors==3.0.10
requests==2.27.1
psycopg2==2.9.5
python-dotenv==0.21.1
